requirements.txt
future==0.18.3
google-cloud-texttospeech==2.14.0